% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{est_helper.G2}
\alias{est_helper.G2}
\title{log-likelihood; mu, covmat by row (upper triangle),  dim >= 2}
\usage{
est_helper.G2(params, method = "tri")
}
\arguments{
\item{params}{length(params) = p + (p+1)*p/2
params[1:p] stores the mean vector:

-    mu_1 mu_2 ... mu_p

params[-(1:p)] stores the triangle elements of the covariance matrix 
(or the Cholesky Decomposition upper triangle matrix v s.t. Cov = v' v)
 
    sig_{1, 1}, sig_{1, 2}, ..., sig_{1, p}
     
    0         , sig_{2,2} , ..., sig_{2, p}
     
    0 ...
     
    0 ...                      , sig_{p, p}}

\item{method}{"tri" or "chole"}
}
\description{
log-likelihood; mu, covmat by row (upper triangle),  dim >= 2
mu, covmat by row (upper triangle), 100 >= dim >= 2

Generally, method == "chole" is more stable (Non-negative Matrix),
while method == "tri" only takes garantee that the matrix is symtric.
}
