% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{f_lower_idx}
\alias{f_lower_idx}
\title{Lower triangle index of a matrix}
\usage{
f_lower_idx(dim_p)
}
\arguments{
\item{dim_p}{dimension of the matrix}
}
\value{
lower triangle index of a matrix
}
\description{
Lower triangle index of a matrix
}
