% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loglikdim2}
\alias{loglikdim2}
\title{log-likelihood; mu, covmat for 2 dim only}
\usage{
loglikdim2(params)
}
\arguments{
\item{params}{parameters for 2 dim case. params = c(mean1, mean2, sig11, sig12, sig22),
hence length(params) = 5}
}
\value{
log-likelihood
}
\description{
log-likelihood; mu, covmat for 2 dim only
}
