% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permsign}
\alias{permsign}
\title{Generate combination of {1, -1} of length n}
\usage{
permsign(n)
}
\arguments{
\item{n}{length of the combination}
}
\value{
a matrix of all possible combinations of {1, -1} of length n
}
\description{
Generate combination of {1, -1} of length n
}
