% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permute_f}
\alias{permute_f}
\title{Density of multivariate folded normal distribution}
\usage{
permute_f(xvec, mu, sigma2)
}
\arguments{
\item{xvec}{a vector of length n}

\item{mu}{mean, a vector of length n}

\item{sigma2}{variance matrix, a matrix of dimension n x n}
}
\value{
density of multivariate folded normal distribution
}
\description{
Density of multivariate folded normal distribution
}
