% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prodbyrow}
\alias{prodbyrow}
\title{pointwise mult-op by row}
\usage{
prodbyrow(mat, vec)
}
\arguments{
\item{mat}{matrix}

\item{vec}{vector}
}
\value{
a matrix of pointwise multiplication of each row of mat by vec
}
\description{
pointwise mult-op by row
}
